+++ app/(route)/main/search/page.tsx
@@ -1,0 +1,5 @@
+const SearchPage = () => {
+  return <div>검색페이지</div>;
+};
+
+export default SearchPage;
