+++ app/_constant/endPoint.ts
@@ -2,4 +2,7 @@ export const END_POINT = {
   ROOT: '/',
   MAIN: '/main',
   SIGN_UP: '/signUp',
+  SEARCH: '/main/search',
+  ACTIVITY: '/main/activity',
+  USER: '/main/user',
 };
